feat(cli): add banner and version header to doctor and update commands
Extract printBanner/resolveBinaryVersion into internal/cli-banner.ts so
init, doctor, and update all share the same implementation. Doctor and update
now show the ASCII art banner + 'OneBrain Doctor/Update  vX.Y.Z  —  /path'
header matching the init command style.
